Use vtex-test-tools instead of jest
