--- node/tests/data/orderFormMocked.ts
@@ -1,9 +1,0 @@
-const orderFormMocked = () => ({
-  salesChannel: '1',
-  storePreferencesData: {
-    countryCode: 'BR',
-  },
-  shippingData: {},
-})
-
-export default orderFormMocked
