--- node/tests/data/postStockMocked.ts
@@ -1,8 +1,0 @@
-const postStockMocked = () => ({
-  'D25133K-B2': 100,
-  '14-556': 5,
-  '123': 100,
-  'DCF880L2-BR': 1515,
-})
-
-export default postStockMocked
