--- node/tests/data/simulationMocked.ts
@@ -1,10 +1,0 @@
-const simulationMocked = () => ({
-  items: [
-    {
-      sku: 'SKU-MOCKED-1',
-      id: '100',
-    },
-  ],
-})
-
-export default simulationMocked
